完成了所有常用表的增删改查接口，新增收藏功能

+++ framework/src/main/java/com/ori/domain/vo/ArticleListVo.java
@@ -47,7 +47,7 @@ public class ArticleListVo {
     /**
      * 评论数量
      */
-    private Long commentCount;
+    private Integer commentCount;
 
     /**
      * 浏览数量
